Integrate project settings into DashboardLive
Move project settings (evogit.toml config, foreign repo management)
from a separate /settings/project page into DashboardLive as an
inline collapsible panel that uses @active_project directly.

Changes:
- dashboard_live.ex: Add project settings toggle button, collapsible
  panel with evogit.toml config display, foreign repos list with
  add/remove functionality. Reset settings panel on project switch.
- settings_live.ex: Remove link to /settings/project page.
- router.ex: Remove /settings/project route.
- layouts.ex: Remove :project_settings from Settings active check.
- Delete project_settings_live.ex (functionality moved to dashboard).
