+++ apps/evo_dash/lib/evo_dash_web/live/dashboard_live.ex
@@ -1,6 +1,7 @@
 defmodule EvoDashWeb.DashboardLive do
   use EvoDashWeb, :live_view
   alias EvoDash.TaskRegistry
+  alias EvoGit.Core.ForeignRepo
 
   @impl true
   def render(assigns) do
@@ -66,6 +67,179 @@ defmodule EvoDashWeb.DashboardLive do
             mode_info={@task_mode_info}
           />
         </div>
+
+        <!-- Project Settings Toggle Button -->
+        <div class="mb-4">
+          <button
+            class="btn btn-sm gap-2"
+            phx-click="toggle_project_settings"
+          >
+            <.icon name="hero-cog-6-tooth" class="size-4" />
+            <%= if @show_project_settings do %>
+              Hide Project Settings
+            <% else %>
+              Project Settings
+            <% end %>
+          </button>
+        </div>
+
+        <%= if @show_project_settings do %>
+          <!-- Project Config Section -->
+          <div class="mb-6 bg-base-100 rounded-2xl shadow-lg border border-base-200 overflow-hidden">
+            <div class="bg-gradient-to-br from-accent/10 via-accent/5 to-transparent p-6">
+              <h2 class="text-lg font-semibold flex items-center gap-2">
+                <.icon name="hero-document-text" class="size-5 text-accent" /> evogit.toml Configuration
+              </h2>
+            </div>
+            <div class="p-6 pt-2">
+              <div class="grid grid-cols-1 sm:grid-cols-2 gap-4">
+                <div class="bg-base-200/40 rounded-lg p-3 border border-base-200">
+                  <p class="text-xs text-base-content/50 font-medium uppercase tracking-wide">Project Root</p>
+                  <p class="text-sm font-mono mt-1">{@active_project}</p>
+                </div>
+                <div class="bg-base-200/40 rounded-lg p-3 border border-base-200">
+                  <p class="text-xs text-base-content/50 font-medium uppercase tracking-wide">Config File</p>
+                  <p class="text-sm mt-1">
+                    <%= if @project_config do %>
+                      <span class="badge badge-success badge-sm gap-1">
+                        <.icon name="hero-check-circle" class="size-3" /> Present
+                      </span>
+                    <% else %>
+                      <span class="badge badge-ghost badge-sm gap-1">
+                        <.icon name="hero-x-circle" class="size-3" /> Not found
+                      </span>
+                    <% end %>
+                  </p>
+                </div>
+                <div class="bg-base-200/40 rounded-lg p-3 border border-base-200 sm:col-span-2">
+                  <p class="text-xs text-base-content/50 font-medium uppercase tracking-wide">Worktree Init Script</p>
+                  <p class="text-sm font-mono mt-1">
+                    <%= if @worktree_script do %>
+                      <span>{@worktree_script}</span>
+                    <% else %>
+                      <span class="text-base-content/30">Not configured</span>
+                    <% end %>
+                  </p>
+                </div>
+              </div>
+            </div>
+          </div>
+
+          <!-- Foreign Repos Section -->
+          <div class="mb-6 bg-base-100 rounded-2xl shadow-lg border border-base-200 overflow-hidden">
+            <div class="bg-gradient-to-br from-secondary/10 via-secondary/5 to-transparent p-6">
+              <h2 class="text-lg font-semibold flex items-center gap-2">
+                <.icon name="hero-server-stack" class="size-5 text-secondary" /> Foreign Repositories
+              </h2>
+            </div>
+            <div class="p-6 pt-2">
+              <%= if @foreign_repos == [] do %>
+                <div class="text-center py-8 text-base-content/40">
+                  <.icon name="hero-folder-minus" class="size-12 mx-auto mb-2 opacity-30" />
+                  <p class="text-sm">No repositories registered</p>
+                </div>
+              <% else %>
+                <div class="space-y-3">
+                  <%= for repo <- @foreign_repos do %>
+                    <div class="flex items-center gap-3 bg-base-200/40 rounded-lg p-3 border border-base-200">
+                      <span class={"badge #{if ForeignRepo.primary?(repo.id), do: "badge-primary", else: "badge-ghost"} badge-sm font-mono"}>
+                        {repo.id}
+                      </span>
+                      <div class="flex-1 min-w-0">
+                        <p class="text-sm font-mono truncate">{repo.root}</p>
+                        <%= if repo.name && repo.name != Atom.to_string(repo.id) do %>
+                          <p class="text-xs text-base-content/50">{repo.name}</p>
+                        <% end %>
+                      </div>
+                      <%= unless ForeignRepo.primary?(repo.id) do %>
+                        <button
+                          class="btn btn-ghost btn-xs text-error"
+                          phx-click="remove_foreign_repo"
+                          phx-value-repo_id={repo.id}
+                        >
+                          <.icon name="hero-trash" class="size-3.5" />
+                        </button>
+                      <% end %>
+                    </div>
+                  <% end %>
+                </div>
+              <% end %>
+            </div>
+          </div>
+
+          <!-- Add Foreign Repo -->
+          <div class="mb-6">
+            <%= if @show_add_foreign_repo_form do %>
+              <div class="bg-base-100 rounded-2xl shadow-lg border border-base-200 overflow-hidden">
+                <div class="bg-gradient-to-br from-success/10 via-success/5 to-transparent p-6">
+                  <h2 class="text-lg font-semibold flex items-center gap-2">
+                    <.icon name="hero-plus-circle" class="size-5 text-success" /> Add Foreign Repository
+                  </h2>
+                </div>
+                <div class="p-6 pt-2">
+                  <.form for={%{}} phx-submit="add_foreign_repo" class="space-y-4">
+                    <div class="grid grid-cols-1 sm:grid-cols-3 gap-4">
+                      <div>
+                        <label class="label">
+                          <span class="label-text text-xs font-medium uppercase tracking-wide">Repo ID</span>
+                        </label>
+                        <input
+                          type="text"
+                          name="repo_id"
+                          value={@new_repo_id}
+                          placeholder="e.g., original"
+                          class="input input-bordered input-sm w-full font-mono"
+                          required
+                        />
+                      </div>
+                      <div>
+                        <label class="label">
+                          <span class="label-text text-xs font-medium uppercase tracking-wide">Path</span>
+                        </label>
+                        <input
+                          type="text"
+                          name="path"
+                          value={@new_repo_path}
+                          placeholder="/absolute/path/to/repo"
+                          class="input input-bordered input-sm w-full font-mono"
+                          required
+                        />
+                      </div>
+                      <div>
+                        <label class="label">
+                          <span class="label-text text-xs font-medium uppercase tracking-wide">Name (optional)</span>
+                        </label>
+                        <input
+                          type="text"
+                          name="name"
+                          value={@new_repo_name}
+                          placeholder="Human-readable name"
+                          class="input input-bordered input-sm w-full"
+                        />
+                      </div>
+                    </div>
+                    <div class="flex gap-2">
+                      <button type="submit" class="btn btn-primary btn-sm gap-2">
+                        <.icon name="hero-plus" class="size-4" /> Add Repository
+                      </button>
+                      <button
+                        type="button"
+                        class="btn btn-ghost btn-sm"
+                        phx-click="toggle_add_foreign_repo_form"
+                      >
+                        Cancel
+                      </button>
+                    </div>
+                  </.form>
+                </div>
+              </div>
+            <% else %>
+              <button class="btn btn-outline btn-sm gap-2" phx-click="toggle_add_foreign_repo_form">
+                <.icon name="hero-plus-circle" class="size-4" /> Add Foreign Repo
+              </button>
+            <% end %>
+          </div>
+        <% end %>
 
       <% else %>
         <!-- No Active Project State -->
@@ -267,6 +441,14 @@ defmodule EvoDashWeb.DashboardLive do
       |> assign(:show_open_project_form, false)
       |> assign(:recent_projects, recent_projects)
       |> assign(:path_suggestions, [])
+      |> assign(:show_project_settings, false)
+      |> assign(:project_config, nil)
+      |> assign(:worktree_script, nil)
+      |> assign(:foreign_repos, [])
+      |> assign(:show_add_foreign_repo_form, false)
+      |> assign(:new_repo_id, "")
+      |> assign(:new_repo_path, "")
+      |> assign(:new_repo_name, "")
       |> assign_form_defaults()
 
     config_status =
@@ -290,6 +472,19 @@ defmodule EvoDashWeb.DashboardLive do
         TaskRegistry.list_tasks_by_path(socket.assigns.active_project)
       else
         TaskRegistry.list_tasks()
+      end
+
+    socket =
+      if socket.assigns.show_project_settings and socket.assigns.active_project do
+        {project_config, worktree_script} = load_project_config(socket.assigns.active_project)
+        foreign_repos = load_foreign_repos()
+
+        socket
+        |> assign(:project_config, project_config)
+        |> assign(:worktree_script, worktree_script)
+        |> assign(:foreign_repos, foreign_repos)
+      else
+        socket
       end
 
     {:noreply, assign(socket, :tasks, new_tasks)}
@@ -317,6 +512,7 @@ defmodule EvoDashWeb.DashboardLive do
        |> assign(:task_mode, mode)
        |> assign(:task_mode_info, mode_info)
        |> assign(:show_open_project_form, false)
+       |> assign(:show_project_settings, false)
        |> assign(:recent_projects, TaskRegistry.list_recent_projects())
        |> put_flash(:info, mode_info)}
     else
@@ -338,7 +534,8 @@ defmodule EvoDashWeb.DashboardLive do
      |> assign(:active_project, path)
      |> assign(:tasks, tasks)
      |> assign(:task_mode, mode)
-     |> assign(:task_mode_info, mode_info)}
+     |> assign(:task_mode_info, mode_info)
+     |> assign(:show_project_settings, false)}
   end
 
   @impl true
@@ -392,6 +589,125 @@ defmodule EvoDashWeb.DashboardLive do
   end
 
   @impl true
+  def handle_event("toggle_project_settings", _params, socket) do
+    show = !socket.assigns.show_project_settings
+
+    socket =
+      if show do
+        {project_config, worktree_script} = load_project_config(socket.assigns.active_project)
+        foreign_repos = load_foreign_repos()
+
+        socket
+        |> assign(:project_config, project_config)
+        |> assign(:worktree_script, worktree_script)
+        |> assign(:foreign_repos, foreign_repos)
+      else
+        socket
+      end
+
+    {:noreply, assign(socket, :show_project_settings, show)}
+  end
+
+  @impl true
+  def handle_event("toggle_add_foreign_repo_form", _params, socket) do
+    {:noreply,
+     socket
+     |> assign(:show_add_foreign_repo_form, !socket.assigns.show_add_foreign_repo_form)
+     |> assign(:new_repo_id, "")
+     |> assign(:new_repo_path, "")
+     |> assign(:new_repo_name, "")}
+  end
+
+  @impl true
+  def handle_event("add_foreign_repo", params, socket) do
+    repo_id_str = String.trim(params["repo_id"] || "")
+    path = String.trim(params["path"] || "")
+    name = String.trim(params["name"] || "")
+
+    cond do
+      repo_id_str == "" ->
+        {:noreply, put_flash(socket, :error, "Repo ID cannot be empty.")}
+
+      path == "" ->
+        {:noreply, put_flash(socket, :error, "Path cannot be empty.")}
+
+      not String.starts_with?(path, "/") ->
+        {:noreply, put_flash(socket, :error, "Path must be absolute (start with /).")}
+
+      true ->
+        repo_id = String.to_atom(repo_id_str)
+
+        repo =
+          if name != "" do
+            ForeignRepo.new(repo_id, path, name: name)
+          else
+            ForeignRepo.new(repo_id, path)
+          end
+
+        try do
+          case EvoGit.AgentScheduler.register_foreign_repo(repo) do
+            :ok ->
+              foreign_repos = load_foreign_repos()
+
+              {:noreply,
+               socket
+               |> assign(:foreign_repos, foreign_repos)
+               |> assign(:show_add_foreign_repo_form, false)
+               |> assign(:new_repo_id, "")
+               |> assign(:new_repo_path, "")
+               |> assign(:new_repo_name, "")
+               |> put_flash(:info, "Foreign repo '#{repo_id_str}' registered successfully.")}
+
+            {:error, {:already_exists, id}} ->
+              {:noreply,
+               socket
+               |> put_flash(:error, "Repo '#{id}' is already registered.")}
+          end
+        rescue
+          e ->
+            {:noreply,
+             socket
+             |> put_flash(:error, "Failed to register repo: #{Exception.message(e)}")}
+        catch
+          _, _ ->
+            {:noreply,
+             put_flash(socket, :error, "Failed to register repo: scheduler not available.")}
+        end
+    end
+  end
+
+  @impl true
+  def handle_event("remove_foreign_repo", %{"repo_id" => repo_id_str}, socket) do
+    repo_id = String.to_atom(repo_id_str)
+
+    try do
+      case EvoGit.AgentScheduler.unregister_foreign_repo(repo_id) do
+        :ok ->
+          foreign_repos = load_foreign_repos()
+
+          {:noreply,
+           socket
+           |> assign(:foreign_repos, foreign_repos)
+           |> put_flash(:info, "Foreign repo '#{repo_id_str}' removed successfully.")}
+
+        {:error, :cannot_unregister_primary} ->
+          {:noreply, put_flash(socket, :error, "Cannot remove the primary repository.")}
+
+        {:error, {:not_found, id}} ->
+          {:noreply, put_flash(socket, :error, "Repo '#{id}' not found.")}
+      end
+    rescue
+      e ->
+        {:noreply,
+         socket
+         |> put_flash(:error, "Failed to remove repo: #{Exception.message(e)}")}
+    catch
+      _, _ ->
+        {:noreply, put_flash(socket, :error, "Failed to remove repo: scheduler not available.")}
+    end
+  end
+
+  @impl true
   def handle_event("path_input", %{"path" => value}, socket) do
     suggestions = path_suggestions(value)
     {:noreply, assign(socket, :path_suggestions, suggestions)}
@@ -440,7 +756,10 @@ defmodule EvoDashWeb.DashboardLive do
         {:ok, task} ->
           {:noreply,
            socket
-           |> put_flash(:info, "#{String.capitalize(to_string(task_type))} task started with ID: #{task.id}")
+           |> put_flash(
+             :info,
+             "#{String.capitalize(to_string(task_type))} task started with ID: #{task.id}"
+           )
            |> assign(:tasks, TaskRegistry.list_tasks_by_path(path))}
 
         {:error, reason} ->
@@ -581,4 +900,39 @@ defmodule EvoDashWeb.DashboardLive do
     end
   end
 
+  # Project Settings Helpers
+
+  defp load_project_config(nil), do: {nil, nil}
+
+  defp load_project_config(project_root) do
+    try do
+      config = EvoGit.ProjectConfig.read(project_root)
+
+      worktree_script =
+        case config do
+          %{"worktree" => %{"script" => script}} when is_binary(script) -> script
+          _ -> nil
+        end
+
+      {config, worktree_script}
+    rescue
+      _ -> {nil, nil}
+    catch
+      _, _ -> {nil, nil}
+    end
+  end
+
+  defp load_foreign_repos do
+    try do
+      repos = EvoGit.AgentScheduler.get_foreign_repos()
+
+      Enum.sort_by(repos, fn repo ->
+        {if(ForeignRepo.primary?(repo.id), do: 0, else: 1), repo.id}
+      end)
+    rescue
+      _ -> []
+    catch
+      _, _ -> []
+    end
+  end
 end
